Integrate scanbeam into beamtable
